Fix external link detection
It was hard-coded to compare to mediocre.com... fixed so it compares to
Mehdown.baseUrl instead. Also removed comparisons for 127.0.0.1 and
localhost, since one of those should be set appropriately in dev
environments as the baseUrl.

This also fixes a bug adding the correct protocol to a scheme-less
domain if the domain represents the baseUrl.

Left in the test for "127.0.0.1 href" by temporarily resetting baseUrl
for that test only. But, changed usernames tests to expect mediocre.com
instead of 127.0.0.1:8000, as that is what is now set as the baseUrl for
tests.

Fixes #18